Update the processing of items for the top menu
The counter in the foreach loop was not being calculated correctly. This
resolved in the 'last' css class not being set on the output when one or
more parent items are not active.

With this commit the items will be filtered first before they are being
processed for the html output.