Make the lib-icon-font to have as a parameter the font-weight
Throught the projects from time to time you need to give a new font weight. The lib-icon-font used to give a normal font-weight hardcoded. Now this is changed with a variable that has the "normal" value as default but can be changed at any moment when you call the function